use-cases/agent-factory-with-subagents/agents/psychology_content_architect/requirements.txt
# Psychology Content Architect Agent Requirements

# Core Pydantic AI framework
pydantic-ai>=0.0.15
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Async HTTP requests
httpx>=0.24.0

# Environment configuration
python-dotenv>=1.0.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Optional: Web scraping for research
beautifulsoup4>=4.12.0
requests>=2.31.0

# Optional: Additional tools for psychology testing
scikit-learn>=1.3.0  # For statistical analysis
matplotlib>=3.7.0    # For visualization (if needed)

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0